pkg/imports.buildRelation -> pkg/logical.BuildRelation
It makes sense for the code that builds relations to be placed in the
logical package, closer to the hardcode-Postgres logic.

Also includes a clean-up of other logical helpers, and clearer comments
about type support.
